+++ examples/basic_table_image.rb
@@ -9,31 +9,31 @@ require 'open-uri'
 
 include Glimmer
 
-IMAGES = []
+IMAGE_ROWS = []
 
-# 50.times do |i|
-#   url = format('https://www.ghibli.jp/gallery/thumb-redturtle%03d.png', (i + 1))
-#   puts "Processing Image: #{url}"
-#   f = URI.open(url)
-#   canvas = ChunkyPNG::Canvas.from_io(f)
-#   f.close
-#   data = canvas.to_rgba_stream
-#   width = canvas.width
-#   height = canvas.height
-#   image = image(width, height) {
-#     image_part(image, data, width, height, width * 4)
-#   }
-#   IMAGES << [image]
-# rescue StandardError => e
-#   warn url, e.message
-# end
+50.times do |i|
+  url = format('https://www.ghibli.jp/gallery/thumb-redturtle%03d.png', (i + 1))
+  puts "Processing Image: #{url}"
+  f = URI.open(url)
+  canvas = ChunkyPNG::Canvas.from_io(f)
+  f.close
+  data = canvas.to_rgba_stream
+  width = canvas.width
+  height = canvas.height
+  img = image(width, height) {
+    image_part(data, width, height, width * 4)
+  }
+  IMAGE_ROWS << [img] # array of one column cell
+rescue StandardError => e
+  warn url, e.message
+end
 
 window('The Red Turtle', 310, 350, false) {
   horizontal_box {
     table {
       image_column('www.ghibli.jp/works/red-turtle', 0)
       
-#       cell_rows IMAGES
+      cell_rows IMAGE_ROWS
     }
     
   }